refactor(mcp-gateway): move gateway modules into their own directory
routes/ had grown 37 mcp-* files, 13 of them one gateway cluster sharing a
`mcp-gateway.` filename prefix that was doing the work a directory should.

They now live in routes/mcp-gateway/, dropping the prefix: index, protocol,
utils, mrtr, and their tests. Pure moves and import rewrites — no behavior
change.

The sibling mcp-app-gateway.utils.ts and mcp-server-gateway.utils.ts stay
put deliberately: despite the name they belong to the app and server proxy
routes, and moving them under the gateway directory would misfile them.
