+++ platform/backend/src/routes/a2a.ts
@@ -10,7 +10,7 @@ import { ProviderError } from "@/routes/chat/errors";
 import {
   extractBearerToken,
   validateMCPGatewayToken,
-} from "@/routes/mcp-gateway.utils";
+} from "@/routes/mcp-gateway/utils";
 import { ApiError, UuidIdSchema } from "@/types";
 
 /**
