+++ platform/backend/src/routes/mcp-gateway/index.ts
@@ -7,13 +7,14 @@ import type { TokenAuthContext } from "@/clients/mcp-client";
 import config from "@/config";
 import { AgentModel, McpToolCallModel } from "@/models";
 import { UuidOrSlugSchema } from "@/types";
+import { getPublicRequestOrigin } from "../request-origin";
 import {
   deriveStatePrincipal,
   extractMrtrParams,
   readClientCapabilities,
   supportsInputRequired,
   verifyRequestState,
-} from "./mcp-gateway.mrtr";
+} from "./mrtr";
 import {
   buildDiscoverResult,
   extractTraceContext,
@@ -26,7 +27,7 @@ import {
   STATELESS_MCP_PROTOCOL_REVISION,
   SUPPORTED_MCP_PROTOCOL_REVISIONS,
   validateRoutingHeaders,
-} from "./mcp-gateway.protocol";
+} from "./protocol";
 import {
   authenticateMCPGatewayRequest,
   createAgentServer,
@@ -37,8 +38,7 @@ import {
   extractPassthroughHeaders,
   extractProfileIdAndTokenFromRequest,
   validateMCPGatewayToken,
-} from "./mcp-gateway.utils";
-import { getPublicRequestOrigin } from "./request-origin";
+} from "./utils";
 
 // =============================================================================
 // MCP Gateway request handling (stateless mode)
