+++ platform/backend/src/routes/mcp-gateway/mrtr.ts
