+++ platform/backend/src/routes/mcp-gateway/protocol.ts
